coffeelog/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
Jinja2==3.1.6
python-dotenv==1.1.1
httpx==0.28.1
google-auth==2.40.3
itsdangerous==2.2.0
requests==2.32.3
